🔧 Use dash in ingress config
Use dash in ingress config to allow for the use of mTLS to hide our
staging apps from the general public. CloudFlare's proxy server by
default only protects us on the root domain and one level down, which
would not be possible with a link like
`staging.api.masterofcubesau.com`. To workaround this, we will use
`staging-` instead.